Finished styling portfolio page, everything looks the way i want it to. Will add functionality in next story.
